✨ feat(file_system): Add the fs doc of ShuoC's cour
We add the file system documentation for Chen Shuo's course in order to
further study the file system. The course analyzes the architecture of ext2
to help novice easily understand FS. The contents we add are as follows:

In the 'file_system' directory:
- 'fs_notes_ShuoChen' now contains individual lessons, ensuring a neat
separation of content.
- 'lesson1_Abstractions.md': Covers abstractions in the file system.
- 'lesson2_Disk_spindle.md': Discusses disk spindle operations.
- 'lesson3_Blocks_and_block_groups.md': Explores the concepts of blocks
and block groups.
- 'lesson4_kernel_file_table.md': Details the kernel file table.
- 'lesson5_inode.md': Focuses on inodes.
- 'lesson6_Directory.md': Covers directory structures.
- 'lesson7_Disk_layout.md': Discusses the layout of the disk.
- 'lesson8_summary.md': Provides a summary of the lessons.

In the 'img' directory, we have provided proper naming and organization
for image files.